Detect and bail on a cancel request
We'll need to do some work here. I think we can track a pseudo-pid for
every client, but it will still be difficult to know which client is
paired to which server. Pgbouncer sets a direct pointer so it's easy to
follow who's talking to who, and it's safe to do this because it's a
single threaded app. We can probably establish a channel in the
core::spawn method to detect a cancelation from a client using a
pseudo-pid/secret, and have it forward or reply a msg with the current
server's startup message/pid/secret. Still seems a bit hacky for a
seemingly fringe feature.