Add support for parse and sync. Decrease buffe. Fix parse bug.
I can't figure out why this impl is so much slower than pgbouncer or
direct connect. Current test is a go app that queries a select * from
table (with 500 rows), and for some reason the read at the top of the
transaction loop (before server is acquired) can take 40ms+. That's
super slow. I found with lib/pq I could close the connection between
queries and that was actually faster. Under wireshark it looks like
pgbouncer is consistently writing 4096 byte packets, but tokio seems to
be writing much more data. I'm not sure if that's a tokio optimization
(or has anything to do with it), but it is different from pgbouncer and
there is a big performance delta.